feat: update default variant button
